requirements.txt
streamlit
google-cloud-storage
google-generativeai
langchain-community
google-genai
google-cloud-bigquery